isolating stdin reader /io data and domain structure
