+++ lib/src/domain/io/repository/io_repository.dart
@@ -1,0 +1,5 @@
+abstract class IoRepository{
+  /// Reads a single line from [stdin] asynchronously.
+  Future<String?> readStdinLine();
+}
+
