+++ lib/src/domain/io/use_case/read_line_from_stdin.dart
@@ -1,0 +1,11 @@
+import 'dart:async';
+
+import 'package:pull_request_coverage/src/domain/io/repository/io_repository.dart';
+
+class ReadLineFromStdin {
+  final IoRepository ioRepository;
+
+  const ReadLineFromStdin(this.ioRepository);
+
+  Future<String?> call() => ioRepository.readStdinLine();
+}
